proof of concept to solve sync between the BP5 and the host.
Initially the host view is writable, until the host connects to the BP5.
After this point the host view is read only and BP5 writes need to call refresh_usbmsdrive() to get the host view in sync
When the connection with the BP5 is closed, the host view becomes writable again
TODO:
call refresh_usbmsdrive() to get the host view in sync.
sync fatfs on connection to get the host changes
Notes:
the dummy command does call refresh_usbmsdrive() to verify the POC