SELF-TEST: reduce spurious scripts after self-test
Old button press waited for for GPIO0 to be zero, and immediately returned.

While uncommon, this could cause
unintended execution of `button.scr`
(or other configured script).

Update to wait for button to be pressed
*AND* released, plus wait a little extra
Also wait for button to be released

Avoids spurious loads of `button.scr` or
other script file from file system after
running self-test, simply because user
button press was not instantaneous.