Avoid future pain
Two changes:
* Do not have default when checking BP_VERSION ... always error is not one of the known board types
* Do no comparisons of BP_BOARD_REVISION outside of a corresponding check for BP_VERSION

Although this is more verbose, it is also more accurate and will avoid
acrobatics to "fit" later board revisions properly into the code.  This is
especially true as the BP_BOARD_REVISION is a numeric value, where
we expect comparisons to support `<=`, `>=`, etc.  -- and those
comparisons would get unwieldy quickly if not anchored to a BP version.

Thinking of this another way:  Conceptually, a board revision has no
meaning without also knowing the the BP version.