updating to address pip changes
now using the default cmake instead of the python one to align to the
operting system's preferred directions. this shouldn't really be an
issue as there's nothing extremely fancy going on in cmake

also moving the docker compose to my personal repo for now, need to talk
with the bp team if they're fine with me actually pushing a real org for
them.