should now build docker correctly, also picotool
with the inclusion of these changes, we address the python weirdness for
cmake as well as we now correct the issue with picotool being unable to
bootstrap. it comes at the cost that if we update picotool to a
different version, we have to update the docker, but that appears to be
a documented issue in their bugs so ... until that's better we're going
to have to stick with that.

also added the .devcontainers for vscode users. this should just "bring
up" a build environment for the vscode system. tested only on linux, but
looking for windows volunteers.

do note that if it fails to build, it's likely that you have a
permissions issue on your current build/ directory and need to clean it
up first.