Use correct format specifiers for unsigned data
e.g.,  `%...u` instead of `%...d` for unsigned values

Also, on a hunch, calculated the maximum buffer usage that might occur in `ui_statusbar_update()`.  This buffer is NOT a queue like the other tx buffers, and there is no blocking behavior (by design) when appending to it.

Unfortunately, the manual calculations shows that the maximum characters used far exceeds the statically allocated buffer size.

STRONGLY recommend to mark unsafe functions like `sprintf()` as deprecated and use safe alternatives.