Remove strings in JSON that are identical to en-us base
At present, it's not known if these strings are intentionally not translated, or whether they have been reviewed and explicitly
chosen to use the en-us version.

As a first step to distinguishing these two states, remove
the identical strings.  Python script is updated to easily show
which strings have not been translated, with instructions to
set explicitly to JSON keywords null if intentionally choosing
to not translate a string.