Much closer
If the row is marked as old (translation not for current EN_US string), and the loaded string was not edited, do not update the stored EN_US string.

If the row has no localized string, or the localized string is an exact match of the EN_US string, then save as NULL string.

Otherwise, save the translation and the EN_US string.